# app/routes/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Designed to be consumed by the Integrated Pipeline Reports (IPR) Client, the endpoints are\navailable for authorized users to consume.",
    "version": "1.0.0",
    "title": "Integrated Pipeline Reports (IPR)",
    "contact": {
      "name": "IPR Team",
      "email": "ipr@bcgsc.ca"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "paths": {
    "/user": {
      "get": {
        "summary": "Get all users",
        "description": "If your account has permissions, this will return all detailed user\nobject descriptions in the database.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "For those authorized, returns all available users in the database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userAssociations"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Denied access to list of all users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ACLDenied"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Authorize new account",
        "description": "Authorize a new user to access IPR. Field 'type' should be equal 'local' or 'bcgsc'.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been successfully added to the system",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userAssociations"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input was provided"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Insufficient access to authorize new accounts"
          },
          "409": {
            "description": "Username is already being used"
          }
        }
      }
    },
    "/user/{user}": {
      "put": {
        "summary": "Update a user account",
        "description": "Update the user specified user details",
        "tags": [
          "User"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "User's UUID string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User account was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "summary": "Remove a single user",
        "description": "Delete a user's access to the API. This does not remove the record from\nthe dataStore the user is located in.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "User's UUID string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The user has been successfully removed"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Insufficient access to remove the user"
          },
          "404": {
            "description": "Unable to find the requested user"
          }
        }
      }
    },
...
    "/user/group": {
      "get": {
        "summary": "Get all groups",
        "description": "Return all groups currently listed in the report ACL database. This is\nnot linked or connected to the BCGSC LDAP database. Changes in one do\nnot affect the other.",
        "tags": [
          "Group",
          "User"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The API will return all available groups it has on record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userGroupAssociations"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "You are not authorized to view this resource"
          }
        }
      },
      "post": {
        "summary": "Create new group",
        "description": "Create a new ACL group in the Report API",
        "tags": [
          "User",
          "Group"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "Name of the new group",
                    "type": "string"
                  },
                  "owner": {
                    "description": "UUID of the owner of the new group",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "owner"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group successfully added to database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userGroupAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "You are not allowed to create new group entries."
          }
        }
      }
    },
    "/user/group/{group}": {
      "get": {
        "summary": "Get a group",
        "description": "Get detailed group object from ident.",
        "tags": [
          "User",
          "Group"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "Group UUID string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group found, API will return the detailed group object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userGroupAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "You are not authorized to view the resource"
          }
        }
      },
      "put": {
        "summary": "Update a group",
        "description": "Update the groups definition in the IPR API database",
        "tags": [
          "User",
          "Group"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "Group UUID string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "Name of the new group",
                    "type": "string"
                  },
                  "owner": {
                    "description": "UUID of the owner of the new group",
                    "type": "string"
                  },
                  "required": [
                    "name",
                    "owner"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userGroupAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "You are not authorized to perform this action"
          },
          "404": {
            "description": "Unable to find the group"
          }
        }
      },
      "delete": {
        "summary": "Remove a group",
        "description": "Remove a group entry from the database. Doing this will immediately\nrevoke group member's access to defined resources.",
        "tags": [
          "User",
          "Group"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "Group UUID string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The group was successfully removed"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "You are not authorized to perform this action"
          },
          "404": {
            "description": "Unable to find the group or user"
          }
        }
      }
    },
...
  },
...
  "components": {
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "ACLDenied": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "example": "You are not authorized to view this resource."
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "headers": {
          "WWW_Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}